Refactor again code per review
