+++ frontend/src/app/(authenticated)/users/delete/page.tsx
@@ -1,45 +1,19 @@
 "use client";
 
 import { useState } from "react";
-import { useRouter } from "next/navigation";
 import Link from "next/link";
-import type { MouseEventHandler } from "react";
 
-import { useAppDispatch, useAppSelector } from "@/stores/hooks";
+import { useAppSelector } from "@/stores/hooks";
 import { BlueButton, RedButton } from "@/components/shared/buttons";
-import { createUserDeleteAction } from "@/features/users/stores/reducers/deleteUserReducer";
 import DeleteAccountModal from "@/features/users/components/deleteAccountModal";
-import { showToast } from "@/stores/toastSlice";
 
 const DeleteUserPage = () => {
-  const router = useRouter();
-  const dispatch = useAppDispatch();
   const authState = useAppSelector((state) => state.auth);
 
   const [isModalOpen, setIsModalOpen] = useState(false);
 
   const handleIsModalOpen = () => {
     setIsModalOpen((state) => !state);
-  };
-
-  const handleDeleteAccount: MouseEventHandler<HTMLButtonElement> = async (
-    event
-  ) => {
-    event.preventDefault();
-    const result = await dispatch(createUserDeleteAction());
-
-    "error" in result
-      ? dispatch(
-          showToast({ text: "アカウントの削除に失敗しました。", type: "error" })
-        )
-      : dispatch(
-          showToast({
-            text: "アカウントが削除されました。ご利用ありがとうございました。",
-            type: "success",
-          })
-        );
-
-    router.push("/");
   };
 
   if (authState.inProgress) {
@@ -59,24 +33,7 @@ const DeleteUserPage = () => {
           <RedButton label="アカウントを削除" onClick={handleIsModalOpen} />
         </div>
       </div>
-      {isModalOpen && (
-        <DeleteAccountModal>
-          <div className="p-6 bg-white rounded-md shadow-md">
-            <h2 className="text-lg font-bold text-center text-gray-700">
-              本当にアカウントを削除しますか？
-            </h2>
-            <p className="mt-2 text-sm text-center text-gray-500">
-              この操作は元に戻すことはできません。
-            </p>
-            <div className="flex justify-center mt-4 space-x-4">
-              <RedButton label="削除する" onClick={handleDeleteAccount} />
-              <Link href={"/todos"}>
-                <BlueButton label="キャンセル" />
-              </Link>
-            </div>
-          </div>
-        </DeleteAccountModal>
-      )}
+      {isModalOpen && <DeleteAccountModal />}
     </div>
   );
 };
