+++ frontend/src/components/shared/modal.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Modal: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+  return (
+    <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50">
+      {children}
+    </div>
+  );
+};
+
+export default Modal;
